change main to webpack exodus
