+++ src/webCache/LocalWebCache.js
@@ -29,11 +29,7 @@ class LocalWebCache extends WebCache {
         }
     }
 
-    async persist(indexHtml) {
-        // extract version from index (e.g. manifest-2.2206.9.json -> 2.2206.9)
-        const version = indexHtml.match(/manifest-([\d\\.]+)\.json/)[1];
-        if(!version) return;
-   
+    async persist(indexHtml, version) {
         const filePath = path.join(this.path, `${version}.html`);
         fs.mkdirSync(this.path, { recursive: true });
         fs.writeFileSync(filePath, indexHtml);
